Add minimize view switcher
