+++ src/components/TableView/index.ts
@@ -1,0 +1,1 @@
+export * from './TableView';
