+++ src/hooks/useTable.tsx
@@ -3,7 +3,7 @@ import { EventBus, FieldType, PanelData } from '@grafana/data';
 import { Button, Icon, useTheme2 } from '@grafana/ui';
 import { ColumnDef } from '@tanstack/react-table';
 import { TestIds } from '../constants';
-import { Styles } from '../styles';
+import { Styles } from '../components/TableView/styles';
 import { Level, PanelOptions, TableItem } from '../types';
 import {
   convertTreeToPlain,
